pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Miyamura80", email = "eitomiyamura@gmail.com" }
]
dependencies = [
    "black>=24.8.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "human-id>=0.2.0",
    "pytest>=8.3.3",
    "termcolor>=2.4.0",
    "loguru>=0.7.3",
    "vulture>=2.14",
    "dspy==3.0.4",
    "langfuse>=2.60.5,<3.0.0",
    "litellm>=1.79.1",
    "tenacity>=9.1.2",
    "pillow>=11.2.1",
    "google-genai>=1.15.0",
    "ty>=0.0.1a9",
    "pytest-env>=1.1.5",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.42",
    "requests>=2.32.4",
    "pytest-asyncio>=1.2.0",
    "fastapi>=0.118.0",
    "PyJWT>=2.8.0",
    "uvicorn>=0.37.0",
    "itsdangerous>=2.2.0",
    "stripe>=13.0.1",
    "workos>=4.0.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.12.0",
    "numpy>=2.3.5",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/python_template"]

[tool.vulture]
exclude = [
    ".venv/",
    ".uv_tools/",
    ".uv_cache/",
    ".uv-tools/",
    ".uv-cache/",
    "tests/**/test_*.py",
    "tests/test_template.py",
    "tests/e2e/e2e_test_base.py",
    "tests/e2e/payments/test_stripe.py",
    "tests/e2e/agent/tools/test_alert_admin.py",
    "utils/llm/",
    "common/",
    "src/utils/logging_config.py",
    "src/utils/context.py",
    "tests/conftest.py",
    "alembic/",
    "src/db/",
    "src/api/routes/",
    "src/api/auth/",
    "src/utils/integration/",
    "src/stripe/",
    "scripts/"
]
ignore_names = [
    "apply_referral",
    "referral_count",
    "get_or_create_referral_code"
]
